Move model to the public API
